0.3.0: Add initial GitHub Actions workflow (#4)
Workflow runs unit and system tests periodically, or on demand. Online documentation states that these specific triggers would only work on the default branch, so use the push trigger for testing on a non-default branch.

Also updating some of the unit tests for consistent behaviour on both Windows and Linux environments.