Improve handling of versenum tags in Web Extractor
Since these tags seem to provide the actual passage number through a tag attribute, it's possible to use that as the tag text instead. This also fixes some edge cases with the NIVUK translation, where specifically searching for an omitted passage could leak passage separator placeholders and make it more difficult to return such passages as a proper list.