Refactor chapter count mapping logic for better flexibility
Previously, the chapter counts were very much assumed to be the same for all translations for a given language, and involved making copies of various functions for each supported language. This has been updated to allow for individual translations to potentially specify what its books are with a single function and also doubles as a way to indicate what language a certain translation is using.