Fix failing unit tests when using newer versions of ruamel.yaml (#17)
Some Web Extractor unit tests were failing because the behaviour of ruamel.yaml has changed in relation to spaces that are the last or first character of a line. Using the existing YAML files (generated using an older ruamel.yaml version) meant that some spaces were doubling up, so this could be a potential issue noticed in actual usage environments.

The system tests were failing for a similar reason, but this time using the JSON Downloader and Extractor seems to be a better fix as it doesn't exhibit this new whitespace handling behaviour.