1.2.0: Add GitHub Actions workflow for publishing to Pypi (#22)
This pull request offloads the complexity of the manual steps related to publishing a new release to Pypi into a GitHub Actions workflow that is able to replicate the same procedure with just a few clicks in the GitHub user interface.

Also pinning twine to a minimum version to ensure some level of stability when run through GitHub Actions.